add perf tests
