+++ Extensions/ArtifactEngine/PerfTests/Perf.ts
@@ -1,0 +1,67 @@
+import * as assert from 'assert';
+import * as path from 'path'
+import * as fs from 'fs'
+
+import * as models from "../Models"
+import * as engine from "../Engine"
+import * as providers from "../Providers"
+
+import { BasicCredentialHandler } from "../Providers/handlers/basiccreds";
+import { PersonalAccessTokenCredentialHandler } from "../Providers/handlers/personalaccesstoken";
+import { ArtifactItemStore } from '../Store/artifactItemStore';
+import { TicketState } from '../Models/ticketState';
+import { ItemType } from '../Models/itemType';
+
+var config = require("./config.json")
+
+describe('perf tests', () => {
+    it('should be able to download large build artifact from vsts drop', function (done) {
+        this.timeout(300000);
+        let processor = new engine.ArtifactEngine();
+
+        let processorOptions = new engine.ArtifactEngineOptions();
+        processorOptions.itemPattern = "**";
+        processorOptions.parallelProcessingLimit = 8;
+        processorOptions.retryIntervalInSeconds = 1000;
+        processorOptions.retryLimit = 5;
+        processorOptions.verbose = true;
+
+        var itemsUrl = "https://testking123.visualstudio.com/_apis/resources/Containers/1902716?itemPath=largedrop&isShallow=false";
+        var variables = {};
+
+        var handler = new PersonalAccessTokenCredentialHandler(config.vsts.pat);
+        var webProvider = new providers.WebProvider(itemsUrl, "vsts.handlebars", variables, handler, { ignoreSslError: false });
+        var dropLocation = path.join(config.dropLocation, "vstsDropWithLargeFiles");
+        var filesystemProvider = new providers.FilesystemProvider(dropLocation);
+
+        processor.processItems(webProvider, filesystemProvider, processorOptions)
+            .then((tickets) => {
+                assert.equal(tickets.filter(x => x.artifactItem.itemType === ItemType.File && x.state === TicketState.Processed).length, 301);
+                done();
+            });
+    });
+
+    it('should be able to download large build artifact from fileshare', function (done) {
+        this.timeout(900000);
+        let processor = new engine.ArtifactEngine();
+
+        let processorOptions = new engine.ArtifactEngineOptions();
+        processorOptions.itemPattern = "**";
+        processorOptions.parallelProcessingLimit = 8;
+        processorOptions.retryIntervalInSeconds = 1000;
+        processorOptions.retryLimit = 5;
+        processorOptions.verbose = true;
+
+        var itemsUrl = "//vscsstor/Users/gykuma/ArtifactEngineTestData/bp/";
+
+        var sourceProvider = new providers.FilesystemProvider(itemsUrl);
+        var dropLocation = path.join(config.dropLocation, "fileshareWithLargeFiles");
+        var filesystemProvider = new providers.FilesystemProvider(dropLocation);
+
+        processor.processItems(sourceProvider, filesystemProvider, processorOptions)
+            .then((tickets) => {
+                assert.equal(tickets.filter(x => x.artifactItem.itemType === ItemType.File && x.state === TicketState.Processed).length, 301);
+                done();
+            });
+    });
+});
